Refactor patch-web-animations-js
